@@ -16,12 +16,10 @@
 package fswalker
 
 import (
-	"context"
 	"crypto/sha256"
 	"encoding/hex"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
@@ -80,8 +78,8 @@ func sha256sum(path string) (string, error) {
 }
 
 // readTextProto reads a text format proto buf and unmarshals it into the provided proto message.
-func readTextProto(ctx context.Context, path string, pb proto.Message) error {
-	b, err := ReadFile(ctx, path)
+func readTextProto(path string, pb proto.Message) error {
+	b, err := os.ReadFile(path)
 	if err != nil {
 		return err
 	}
@@ -89,24 +87,9 @@ func readTextProto(ctx context.Context, path string, pb proto.Message) error {
 }
 
 // writeTextProto writes a text format proto buf for the provided proto message.
-func writeTextProto(ctx context.Context, path string, pb proto.Message) error {
+func writeTextProto(path string, pb proto.Message) error {
 	blob := prototext.Format(pb)
 	// replace message boundary characters as curly braces look nicer (both is fine to parse)
 	blob = strings.Replace(strings.Replace(blob, "<", "{", -1), ">", "}", -1)
-	return WriteFile(ctx, path, []byte(blob), 0644)
-}
-
-// ReadFile reads the file named by filename and returns the contents.
-var ReadFile = func(_ context.Context, filename string) ([]byte, error) {
-	return ioutil.ReadFile(filename)
-}
-
-// WriteFile writes data to a file named by filename.
-var WriteFile = func(_ context.Context, filename string, data []byte, perm os.FileMode) error {
-	return ioutil.WriteFile(filename, data, perm)
-}
-
-// Glob returns the names of all files matching pattern or nil if there is no matching file.
-var Glob = func(_ context.Context, pattern string) ([]string, error) {
-	return filepath.Glob(pattern)
+	return os.WriteFile(path, []byte(blob), 0644)
 }
